Add files via upload
Code to scan Oracle Schema(s) existing indexes and recommend settings for key compression (prefix compression), or alternatively (if specified) implement recommendations. 

This code implements a feature of the Oracle database freely available in Enterprise Edition and will reduce index size (leaf blocks) and increase efficiency (more data per block means more data in the buffer cache which translates into less IO). 

The code requires a table to be created in order to be run (definition included within the package spec). Index key compression would not be recommended in special situations such as a highly concurrent or extremely active throughput system (tables with the index on it). 

The code can be run in various modes, and is recommended to be run only against non-production databases, assess the results and use those results to plan modifications to production. The code will lock tables when creating indexes so if you do run the code in implement mode please be aware of that, especially for very large indexes.

The table created to facilitate this process will log the leaf blocks before and after as well as the original DDL. Ideally this table and code would be deployed against a privileged user with a DBA role granted to it. It could be deployed into the schema to be analyzed or into a more generic one to run against multiple schemas.